backend/requirements.txt
fastapi
uvicorn
pandas
chromadb
sentence-transformers
python-dotenv
faker
requests
pytest
